requirements/api.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn
mlflow-skinny